Change Go version detection in the test workflow
The `actions/setup-go` action has [ability to detect Go version from `go.mod` file][1].
In that case `go-version-file` input needs to be used instead of `go-version`.

As long as there's no need in multiple Go versions (like with the matrix testing) or the SemVer range syntax (like `>=1.17.0`),
`go-version-file` input might be a better choice because it makes `go.mod` file a single source of truth and simplifies maintenance.

[1]: https://github.com/actions/setup-go?tab=readme-ov-file#getting-go-version-from-the-gomod-file