Fix communication issue with Sublime Text
The LSP client of Sublime Text doesn't like the output from an LSP server when it contains something that isn't following the LSP protocol.
It throws an error and stops trying to connect.

To fix the issue, the logline that wasn't following the LSP protocol was replaced with a `window/logMessage` notification.
That allows to still have the message in the output and doesn't break the Sublime LSP client at the same time.

NOTE: The problematic message was published by a logger that wasn't used anywhere else.
      Therefore, the logger was removed as well as the `log-path` command-line optiion related to it.